requirements.txt
mysql.connector
BeautifulTable
pymongo